Task-2567. Refactor API Server to Typescript.
We have refactored all of the routers in the api-server to typescript. Because of this, the endpoint webLogin has to have a consitent type for the data returned to the clent, specifically the jsonStatus.status variable will only return a number or undefined, whereas previously it could potentially return a string, a boolean, a number or undefined. -1, which is an impossible value for a user's verify step will be used in place of false.